# apps/api/openapi/moduflow-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/permissions": {
      "post": {
        "operationId": "PermissionsController_create",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionResponse"
                }
              }
            }
          }
        },
        "summary": "Create a permission",
        "tags": [
          "Permissions"
        ]
      },
      "get": {
        "operationId": "PermissionsController_findAll",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PermissionResponse"
                  }
                }
              }
            }
          }
        },
        "summary": "List permissions",
        "tags": [
          "Permissions"
        ]
      }
    },
    "/api/permissions/{id}": {
      "get": {
        "operationId": "PermissionsController_findOne",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the permission.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionResponse"
                }
              }
            }
          }
        },
        "summary": "Get a permission by id",
        "tags": [
          "Permissions"
        ]
      },
      "patch": {
        "operationId": "PermissionsController_update",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the permission.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionResponse"
                }
              }
            }
          }
        },
        "summary": "Update a permission",
        "tags": [
          "Permissions"
        ]
      },
      "delete": {
        "operationId": "PermissionsController_remove",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the permission.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionResponse"
                }
              }
            }
          }
        },
        "summary": "Delete a permission",
        "tags": [
          "Permissions"
        ]
      }
    },
...
    "/api/roles/{roleId}/permissions/{permissionId}": {
      "delete": {
        "operationId": "RolePermissionsController_removePermission",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the role.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "permissionId",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the permission.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RolePermissionResponse"
                }
              }
            }
          }
        },
        "summary": "Remove a permission from a role",
        "tags": [
          "Role Permissions"
        ]
      }
    },
    "/api/organizations": {
      "post": {
        "operationId": "OrganizationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          }
        },
        "summary": "Create an organization",
        "tags": [
          "Organizations"
        ]
      },
      "get": {
        "operationId": "OrganizationsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationResponse"
                  }
                }
              }
            }
          }
        },
        "summary": "List organizations",
        "tags": [
          "Organizations"
        ]
      }
    },
    "/api/organizations/{id}": {
      "get": {
        "operationId": "OrganizationsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the organization.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          }
        },
        "summary": "Get an organization by id",
        "tags": [
          "Organizations"
        ]
      },
      "patch": {
        "operationId": "OrganizationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the organization.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          }
        },
        "summary": "Update an organization",
        "tags": [
          "Organizations"
        ]
      },
      "delete": {
        "operationId": "OrganizationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the organization.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          }
        },
        "summary": "Delete an organization",
        "tags": [
          "Organizations"
        ]
      }
    }
  },
  "info": {
    "title": "Moduflow API",
    "description": "Moduflow API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "OrganizationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "name": {
            "type": "string",
            "example": "Northwind Homes"
          },
          "slug": {
            "type": "string",
            "example": "northwind-homes"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
...
      "CreatePermissionDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "Permission key.",
            "example": "users.manage",
            "maxLength": 255
          },
          "description": {
            "type": "string",
            "description": "Permission description.",
            "example": "Allows managing users."
          }
        },
        "required": [
          "key"
        ]
      },
      "PermissionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "key": {
            "type": "string",
            "example": "users.manage"
          },
          "description": {
            "type": "object",
            "example": "Allows managing users.",
            "nullable": true
          }
        },
        "required": [
          "id",
          "key",
          "description"
        ]
      },
      "UpdatePermissionDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "Permission key.",
            "example": "users.manage",
            "maxLength": 255
          },
          "description": {
            "type": "string",
            "description": "Permission description.",
            "example": "Allows managing users."
          }
        }
      },
...
      "RolePermissionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "key": {
            "type": "string",
            "example": "users.manage"
          },
          "description": {
            "type": "object",
            "example": "Allows managing users.",
            "nullable": true
          }
        },
        "required": [
          "id",
          "key",
          "description"
        ]
      },
      "CreateOrganizationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Organization name.",
            "example": "Northwind Homes",
            "maxLength": 120
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly unique organization slug.",
            "example": "northwind-homes",
            "maxLength": 80,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
      "UpdateOrganizationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Organization name.",
            "example": "Northwind Homes",
            "maxLength": 120
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly unique organization slug.",
            "example": "northwind-homes",
            "maxLength": 80,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          }
        }
      }
    }
  }
}